// server/src/core/enums/status.enum.ts
export enum Status {    
    OK = 'OK',
    CREATED = 'CREATED',
    ACCEPTED = 'ACCEPTED',
    NO_CONTENT = 'NO CONTENT',
    BAD_REQUEST = 'BAD REQUEST',
    UNAUTHORIZED = 'UNAUTHORIZED',
    FORBIDDEN = 'FORBIDDEN',
    NOT_FOUND = 'NOT FOUND',
    METHOD_NOT_ALLOWED = 'METHOD NOT ALLOWED',
    CONFLICT = 'CONFLICT',
    INTERNAL_SERVER_ERROR = 'INTERNAL SERVER ERROR',
    NOT_IMPLEMENTED = 'NOT IMPLEMENTED',
    BAD_GATEWAY = 'BAD GATEWAY',
    SERVICE_UNAVAILABLE = 'SERVICE UNAVAILABLE'
}